requirements.txt
kivy
kivymd
torch
torchvision
pillow
numpy
gtts
